roast create, update; taste index

+++ database/migrations/2016_10_10_065242_create_brew_readings_table.php
@@ -17,7 +17,7 @@ class CreateBrewReadingsTable extends Migration
             $table->increments('id');
             $table->integer('brew_id')->unsigned()->index();
             $table->double('temperature');
-            $table->integer('time');
+            $table->double('time');
             $table->timestamps();
 
             $table->foreign('brew_id')
